Separating Outcomes Requests into OSOutcomesRequests
