+++ iOS_SDK/OneSignalSDK/OneSignalCore/Source/OneSignalCore.h
@@ -15,24 +15,14 @@
 #import "OneSignalLog.h"
 #import "NSURL+OneSignal.h"
 #import "NSString+OneSignal.h"
+#import "OSRequests.h"
 #import "OneSignalRequest.h"
 #import "OneSignalClient.h"
 #import "OneSignalCoreHelper.h"
 #import "OneSignalTrackFirebaseAnalytics.h"
+#import "OSMacros.h"
 
 @interface OneSignalCore : NSObject
 
 @end
-// Defines let and var in Objective-c for shorter code
-// __auto_type is compatible with Xcode 8+
-#if defined(__cplusplus)
-#define let auto const
-#else
-#define let const __auto_type
-#endif
 
-#if defined(__cplusplus)
-#define var auto
-#else
-#define var __auto_type
-#endif
