+++ iOS_SDK/OneSignalSDK/OneSignalOutcomes/OneSignalOutcomes.h
@@ -8,12 +8,8 @@
 
 #import <Foundation/Foundation.h>
 
-//! Project version number for OneSignalOutcomes.
-FOUNDATION_EXPORT double OneSignalOutcomesVersionNumber;
-
-//! Project version string for OneSignalOutcomes.
-FOUNDATION_EXPORT const unsigned char OneSignalOutcomesVersionString[];
-
+#import "OSOutcomesRequests.h"
+#import "OSSessionManager.h"
 // In this header, you should import all the public headers of your framework using statements like #import <OneSignalOutcomes/PublicHeader.h>
