+++ iOS_SDK/OneSignalSDK/OneSignalOutcomes/Source/Influence/OSInfluenceDataRepository.h
@@ -25,7 +25,6 @@ OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 THE SOFTWARE.
 */
 
-#import "OneSignal.h"
 
 #ifndef OSInfluenceDataRepository_h
 #define OSInfluenceDataRepository_h
