+++ iOS_SDK/OneSignalSDK/OneSignalOutcomes/Source/OutcomeEvents/Controller/V2/OSOutcomeEventsV2Repository.m
@@ -28,8 +28,7 @@ THE SOFTWARE.
 #import <Foundation/Foundation.h>
 
 #import "OSOutcomeEventsV2Repository.h"
-#import "Requests.h"
-#import "OneSignalClient.h"
+#import <OneSignalCore/OneSignalCore.h>
 #import "OSOutcomeEvent.h"
 
 @implementation OSOutcomeEventsV2Repository
