+++ iOS_SDK/OneSignalSDK/OneSignalOutcomes/Source/OutcomeEvents/Model/V1/OSCachedUniqueOutcome.m
@@ -27,7 +27,7 @@
 
 #import <Foundation/Foundation.h>
 #import "OSCachedUniqueOutcome.h"
-#import "OneSignalCommonDefines.h"
+#import <OneSignalCore/OneSignalCore.h>
 
 @implementation OSCachedUniqueOutcome
