+++ iOS_SDK/OneSignalSDK/Source/OneSignal.h
@@ -44,7 +44,7 @@
 #import <Foundation/Foundation.h>
 #import <UserNotifications/UserNotifications.h>
 #import <OneSignalCore/OneSignalCore.h>
-
+#import <OneSignalOutcomes/OneSignalOutcomes.h>
 #pragma clang diagnostic push
 #pragma clang diagnostic ignored "-Wstrict-prototypes"
 #pragma clang diagnostic ignored "-Wnullability-completeness"
@@ -236,40 +236,40 @@ typedef NS_ENUM(NSUInteger, OSNotificationActionType)  {
 - (void)onDidDismissInAppMessage:(OSInAppMessage *)message;
 @end
 
-/* OneSignal Influence Types */
-typedef NS_ENUM(NSUInteger, OSInfluenceType) {
-    DIRECT,
-    INDIRECT,
-    UNATTRIBUTED,
-    DISABLED
-};
-/* OneSignal Influence Channels */
-typedef NS_ENUM(NSUInteger, OSInfluenceChannel) {
-    IN_APP_MESSAGE,
-    NOTIFICATION,
-};
+///* OneSignal Influence Types */
+//typedef NS_ENUM(NSUInteger, OSInfluenceType) {
+//    DIRECT,
+//    INDIRECT,
+//    UNATTRIBUTED,
+//    DISABLED
+//};
+///* OneSignal Influence Channels */
+//typedef NS_ENUM(NSUInteger, OSInfluenceChannel) {
+//    IN_APP_MESSAGE,
+//    NOTIFICATION,
+//};
 
-@interface OSOutcomeEvent : NSObject
-
-// Session enum (DIRECT, INDIRECT, UNATTRIBUTED, or DISABLED) to determine code route and request params
-@property (nonatomic) OSInfluenceType session;
-
-// Notification ids for the current session
-@property (strong, nonatomic, nullable) NSArray *notificationIds;
-
-// Id or name of the event
-@property (strong, nonatomic, nonnull) NSString *name;
-
-// Time of the event occurring
-@property (strong, nonatomic, nonnull) NSNumber *timestamp;
-
-// A weight to attach to the outcome name
-@property (strong, nonatomic, nonnull) NSDecimalNumber *weight;
-
-// Convert the object into a NSDictionary
-- (NSDictionary * _Nonnull)jsonRepresentation;
-
-@end
+//@interface OSOutcomeEvent : NSObject
+//
+//// Session enum (DIRECT, INDIRECT, UNATTRIBUTED, or DISABLED) to determine code route and request params
+//@property (nonatomic) OSInfluenceType session;
+//
+//// Notification ids for the current session
+//@property (strong, nonatomic, nullable) NSArray *notificationIds;
+//
+//// Id or name of the event
+//@property (strong, nonatomic, nonnull) NSString *name;
+//
+//// Time of the event occurring
+//@property (strong, nonatomic, nonnull) NSNumber *timestamp;
+//
+//// A weight to attach to the outcome name
+//@property (strong, nonatomic, nonnull) NSDecimalNumber *weight;
+//
+//// Convert the object into a NSDictionary
+//- (NSDictionary * _Nonnull)jsonRepresentation;
+//
+//@end
 
 typedef NS_ENUM(NSInteger, OSNotificationPermission) {
     // The user has not yet made a choice regarding whether your app can show notifications.
@@ -435,8 +435,8 @@ typedef void (^OSWebOpenURLResultBlock)(BOOL shouldOpen);
 typedef void (^OSResultSuccessBlock)(NSDictionary* result);
 typedef void (^OSFailureBlock)(NSError* error);
 
-/*Block for handling outcome event being sent successfully*/
-typedef void (^OSSendOutcomeSuccess)(OSOutcomeEvent* outcome);
+///*Block for handling outcome event being sent successfully*/
+//typedef void (^OSSendOutcomeSuccess)(OSOutcomeEvent* outcome);
 
 // ======= OneSignal Class Interface =========
 @interface OneSignal : NSObject
